+++ trilogy-llvm/core/trilogy_set.c
@@ -41,7 +41,7 @@ trilogy_set_value* trilogy_set_assume(trilogy_value* val) {
 void trilogy_set_destroy(trilogy_set_value* set) {
     if (--set->rc == 0) {
         if (set->contents == NULL) return;
-        for (uint64_t i = 0; i < set->len; ++i) {
+        for (size_t i = 0; i < set->len; ++i) {
             trilogy_value_destroy(&set->contents[i]);
         }
         free(set->contents);
