+++ trilogy-llvm/core/trilogy_string.h
@@ -5,7 +5,7 @@
 trilogy_string_value*
 trilogy_string_init(trilogy_value* tv, trilogy_string_value* str);
 trilogy_string_value*
-trilogy_string_init_new(trilogy_value* tv, uint64_t len, char* s);
+trilogy_string_init_new(trilogy_value* tv, size_t len, char* s);
 trilogy_string_value*
 trilogy_string_clone_into(trilogy_value* tv, const trilogy_string_value* orig);
 trilogy_string_value* trilogy_string_init_from_c(trilogy_value* tv, char* s);
@@ -17,8 +17,8 @@ void trilogy_string_destroy(trilogy_string_value* val);
 
 char* trilogy_string_as_c(trilogy_string_value* val);
 
-uint64_t trilogy_string_len(trilogy_string_value* val);
-uint32_t trilogy_string_at(trilogy_string_value* str, uint64_t index);
+size_t trilogy_string_len(trilogy_string_value* val);
+uint32_t trilogy_string_at(trilogy_string_value* str, size_t index);
 trilogy_string_value* trilogy_string_concat(
     trilogy_value* rt, trilogy_string_value* lhs, trilogy_string_value* rhs
 );
